+++ psd-web/app/controllers/account_name_controller.rb
@@ -1,0 +1,18 @@
+class AccountNameController < ApplicationController
+  def show
+    @user = current_user
+    render "account/name"
+  end
+
+  def update
+    @user = current_user
+
+    @user.name = params.dig(:user, :name)
+
+    if @user.save(context: :change_name)
+      redirect_to account_path
+    else
+      render "account/name"
+    end
+  end
+end