+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## 2020-04-09
+## 2020-04-16
 - Extended the question to identify cases related to the coronavirus outbreak to:
   - Allegations
   - Enquiries
   - Projects
+
+## 2020-04-15
+- Added Your account page and ability to change your name
 
 ## 2020-04-07
 - Added ability to edit the coronavirus related flag